Added globals, more tests
Added `$scope` properties (globals as far as the game is concerned) for
`citizens` and `employed`. `citizens` is the total number of people on
the board. `employed` is the total number of people with jobs on the
board.

Added tests for `citizens` and `employed`. The tests don't (and can't)
test in the context of the `$scope.random_reverse` method I created.
`random_reverse` exists to break the left-to-right-top-to-bottom path of
the residential processor. The `random_revarse` method only executes as
part of the `$interval` loop, so it is outside the scope of tests.