Got new residential tile type working
I've re-thought the way tiles get created.

Basically, I've stripped all the creation options down to a single
property for residential tiles: beds. A residential tile may have 1 - 7
beds. When there is an even number of beds, each bed hold one person.
When there is an off number of beds one of the beds (the master bed)
holds two people. The master bed takes up an entire floor and so has the
room for an extra person.

The idea right now is that all tile types will work similarly to this.
Other tile types may have two properties, but I'm thinking I'll kep it
to one for now.